requirements.txt
aiohttp==3.9.5
APScheduler==3.10.4
Flask==3.1.2
Flask_Cors==4.0.0
google_api_python_client==2.187.0
groq==1.0.0
hypercorn==0.18.0
langdetect==1.0.9
Pillow==12.1.0
praw==7.7.1
protobuf==6.33.4
pystray==0.19.5
python-dotenv==1.2.1
pytz==2024.1
quart==0.20.0
quart_cors==0.8.0
Requests==2.32.5
SQLAlchemy==2.0.30
telethon==1.42.0
tweepy==4.14.0
